Fixed settings

+++ frontend/src/lib/api.js
@@ -11,5 +11,5 @@ export const resetPassword = async ({ verificationCode, password }) =>
   API.post("/auth/password/reset", { verificationCode, password });
 
 export const getUser = async () => API.get("/user");
-export const getSessions = async () => API.get("/sessions");
-export const deleteSession = async (id) => API.delete(`/sessions/${id}`);
+export const getSessions = async () => API.get("/admin/sessions");
+export const deleteSession = async (id) => API.delete(`/admin/sessions/${id}`);
